% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_GUANO.R
\name{guano_md}
\alias{guano_md}
\title{Read GUANO metadata in audio file}
\usage{
guano_md(file)
}
\arguments{
\item{file}{Path to a wav file}
}
\value{
list of named metadata fields
}
\description{
Read GUANO metadata in audio file
}
