% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadata}
\alias{metadata}
\alias{metadata.threshold_detection}
\alias{metadata.blob_detection}
\title{Extract metadata}
\usage{
metadata(x)

\method{metadata}{threshold_detection}(x)

\method{metadata}{blob_detection}(x)
}
\arguments{
\item{x}{an object for which metadata will be extracted}
}
\description{
Extract metadata
}
