% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_zc.R
\name{metadata_zc}
\alias{metadata_zc}
\title{Extract metadata from Zero-Crossing files}
\usage{
metadata_zc(x)
}
\arguments{
\item{x}{an object for which metadata will be extracted}
}
\description{
Extract metadata from Zero-Crossing files
}
