% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myotis-data.R, R/zc-data.R
\docType{data}
\name{myotis}
\alias{myotis}
\alias{zc}
\title{Audio recording of myotis species from United-Kingdom}
\format{
\link[tuneR]{Wave} object

Zero-Crossing object
}
\usage{
myotis

zc
}
\description{
The myotis dataset is a Wave file of 19.73 seconds, 16 bits, mono, 10x time expanded recording with a sampling rate at 50000 Hz.
It contains 20 echolocation calls of several species from the Myotis genus.
The recording was made in United-Kingdom with a D500X bat detector from Pettersson Elektronik AB.

The zc dataset is a Zero-Crossing file of 16384 dots containing a sequence of 24 echolocation calls of a hoary bat (Lasiurus cinereus).
This ZC recording was made in Gatineau Park, Quebec, eastern Canada, during the summer 2017 with a Walkabout bat detector from Titley Scientific.
}
\concept{bat}
\concept{echolocation calls}
\concept{hoary bat}
\concept{myotis}
\concept{zero-crossing}
\keyword{datasets}
