% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_audio.R
\name{read_mp3}
\alias{read_mp3}
\title{Read MP3 files}
\usage{
read_mp3(file, time_exp = 1, ...)
}
\arguments{
\item{file}{a MP3 file.}

\item{time_exp}{integer. Time expansion factor of the recording.
Set to 1 for real-time recording or above for time expanded recording. Default setting is 1.}

\item{...}{currently not implemented.}
}
\value{
A \link[tuneR]{Wave} object.
}
\description{
A thin wrapped around \link[tuneR]{readMP3} from the package tuneR.
}
\examples{
\donttest{
filepath <- system.file("extdata", "recording.mp3", package = "bioacoustics")
read_mp3(filepath)
}

}
