% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_zc.R
\name{read_zc}
\alias{read_zc}
\title{Read Zero-Crossing files}
\usage{
read_zc(file)
}
\arguments{
\item{file}{a Zero-Crossing file.}
}
\value{
an object of class 'zc'.
}
\description{
Read Zero-Crossing files (.zc, .#) from various bat recorders
}
\examples{
\dontrun{
zc <- read_zc("file")
}

}
