% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{resample}
\alias{resample}
\title{Resample}
\usage{
resample(wave, to)
}
\arguments{
\item{wave}{a \link[tuneR]{Wave} object.}

\item{to}{target frequency in Hz}
}
\value{
a \link[tuneR]{Wave} object.
}
\description{
Resample a Wave object to a given sampling frequency.
}
\examples{
data(myotis)
myotis_192 <- resample(myotis, to = 192000)
spectro(myotis_192, tlim = c(1, 1.5))

}
