% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{to_dB}
\alias{to_dB}
\title{Convert to dB}
\usage{
to_dB(x, ref = 1)
}
\arguments{
\item{x}{numeric. Vector of amplitude values (V1).}

\item{ref}{numeric. Reference value (V0) to calculate the ratio (V1/V0).}
}
\description{
Convert amplitude to decibel (dB) values
}
\keyword{internal}
