% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_zc.R
\name{write_zc}
\alias{write_zc}
\title{Write Zero-Crossing files}
\usage{
write_zc(zc, filename)
}
\arguments{
\item{zc}{an object of class 'zc'.}

\item{filename}{path or connection to write.}
}
\description{
Write Zero-Crossing files (.zc, .#)
}
\examples{
data(zc)
filename <- tempfile()
write_zc(zc, filename = filename)

}
