% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadata}
\alias{metadata}
\alias{metadata.character}
\alias{metadata.blob_detection}
\alias{metadata.threshold_detection}
\alias{metadata.zc}
\alias{metadata.Wave}
\title{Extract metadata}
\usage{
metadata(x, ...)

\method{metadata}{character}(x, file_type = c(file_type_guess(x), "wav", "zc"), ...)

\method{metadata}{blob_detection}(x, ...)

\method{metadata}{threshold_detection}(x, ...)

\method{metadata}{zc}(x, ...)

\method{metadata}{Wave}(x, ...)
}
\arguments{
\item{x}{an object for which metadata will be extracted}

\item{...}{further arguments passed to or from other methods.}

\item{file_type}{type of file to read metadata from. Wav and Zero-Crossing files are currently supported.}
}
\description{
Extract metadata

Extract metadata from Zero-Crossing files

Extract metadata from a Wave object
}
